requirements.txt
websocket-client==1.6.4
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
cryptography==42.0.5